Add attributes for RFC 5780 (NAT behaviour)
This commit adds two attributes needed for the implementation of RFC
5780 (NAT Behavior Discovery Using STUN). The OTHER-ADDRESS attribute is
returned by a compatable STUN server and includes as its value the
server's second address. The CHANGE-REQUEST attribute is sent by a
client and contains as its value a 32-bit integer that encodes two
flags: one to indicate that the server should respond to the request
from a different IP address, and the other to request a change of port.